% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{basic_parameters_site}
\alias{basic_parameters_site}
\title{Get malariasimple basic parameters for site}
\usage{
basic_parameters_site(
  site,
  start_date = "2000-01-01",
  end_date = "2020-12-31",
  ...
)
}
\arguments{
\item{site}{site object}

\item{start_date}{Start date of simulation in "YYYY-MM-DD" format. 1st January 2000.}

\item{end_date}{End date of simulation in "YYYY-MM-DD" format. Default is 31st December 2020.}

\item{...}{Further arguments to malariasimple::get_parameters()}
}
\description{
Sets up basic malariasimple parameters for a specific site. Equivalent to malariasimple::get_parameters().
Population is taken from site-files and is set to the mean average over the defined simulation years.
}
