% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_production.R
\name{bplotify}
\alias{bplotify}
\title{Convert to form for use with bayesplot package}
\usage{
bplotify(pars_array)
}
\arguments{
\item{pars_array}{3D array in the form [parameter, iteration, chain]. Output from monty_sample()$pars}
}
\description{
Convert to form for use with bayesplot package
}
