% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{carrying_capacity_icl}
\alias{carrying_capacity_icl}
\title{Set carrying capacity according to Imperial College London model}
\usage{
carrying_capacity_icl(rainfall_df, tau, zsat, zmax)
}
\arguments{
\item{rainfall_df}{DataFrame containing columns 'Date' and 'Rainfall'. Output from get_climate_data(.., simplify_outputs = TRUE)}

\item{tau}{Smoothing factor. Larger values indicate greater smoothing and generally lead to a longer mosquito season}

\item{zsat}{Carrying capacity saturation. Larger values of zsat limit maximum carrying capacity.}

\item{zmax}{Washout effects}
}
\description{
Set carrying capacity according to Imperial College London model
}
