% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{carrying_capacity_white}
\alias{carrying_capacity_white}
\title{Set carrying capacity according to White et al. 2011}
\usage{
carrying_capacity_white(rainfall_df, tau)
}
\arguments{
\item{rainfall_df}{DataFrame with containing columns 'Date' and 'Rainfall'. Output from get_climate_data(.., simplify_outputs = TRUE)}

\item{tau}{smoothing factor. Larger values indicate greater smoothing.#}
}
\description{
Carrying capacity proportional to past rainfall weighted by an exponential distribution
}
