% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_production.R
\name{codafy}
\alias{codafy}
\title{Convert to form for use with coda package}
\usage{
codafy(pars_array)
}
\arguments{
\item{pars_array}{3D array in the form [parameter, iteration, chain]. Output from monty_sample()$pars}
}
\description{
Convert to form for use with coda package
}
