% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_samples_list}
\alias{example_samples_list}
\title{An example list of MCMC outputs}
\format{
A list of four monty output objects
}
\usage{
example_samples_list
}
\description{
An example list of MCMC outputs
}
\keyword{datasets}
