% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{filter_site_interventions}
\alias{filter_site_interventions}
\title{Filter sites by interventions usage}
\usage{
filter_site_interventions(
  site_list,
  IRS = NULL,
  ITN = NULL,
  SMC = NULL,
  minimum_prop = 0.05
)
}
\arguments{
\item{IRS}{Should IRS be present?}

\item{ITN}{Should ITN be present?}

\item{SMC}{Should SMC be present?}

\item{minimum_prop}{Minimum coverage at which intervention is considered 'present' (maximum value)}
}
\description{
For each intervention, NULL (default) indicates ignore,
TRUE indicates intervention must be present,
FALSE indicates intervention must not be present
}
