% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{get_climate_data}
\alias{get_climate_data}
\title{Get climate satellite data for a site}
\usage{
get_climate_data(
  site,
  name_2 = NULL,
  data_type = c("rainfall", "temperature"),
  simplify_outputs = FALSE,
  data_directory =
    "//wpia-hn.hpc.dide.ic.ac.uk/vimc-cc/DataCentre/observation/countries/"
)
}
\arguments{
\item{site}{Site object}

\item{name_2}{Name of admin 2 region. If unspecified, climate data for admin 1 will be returned.}

\item{data_type}{A vector of required types of data. "rainfall" and "temperature" are accepted. Other data types must be given by a specific title e.g. "ERA5_Land_specific_humidity"}

\item{simplify_outputs}{If TRUE, outputs will be simplified to just date and unweighted data}

\item{data_directory}{Directory where the climate data is located}
}
\description{
Returns daily population-weighted CHIRPS total rainfall for data_type = "rainfall".
Returns daily ERA5 mean 2m temperature when data_type = "temperature"
}
