% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_production.R
\name{get_dens_plot}
\alias{get_dens_plot}
\title{Get density plot of parameters}
\usage{
get_dens_plot(
  pars_array,
  pars = NULL,
  title = "Parameter Posterior Density",
  burn_in = 0,
  true_vals_df = NULL,
  priors = NULL,
  save_file = NULL,
  height = 20,
  width = 30
)
}
\arguments{
\item{pars}{Names of parameters to show. Default = all parameters}

\item{title}{Plot title}

\item{burn_in}{Length of 'burn-in' to discard from plot}

\item{true_vals_df}{Dataframe with two named columns: variable and value}

\item{priors}{Named list of functions. Each element should be a function taking a numeric vector x and returning the prior density at x. Names must match parameter names.}

\item{save_file}{File location for ggsave}

\item{height}{Height of saved plot in cm}

\item{width}{Width of saved plot in cm}

\item{mcmc_output}{Output from mcmc_go() or monty::monty_sample()}
}
\description{
Get density plot of parameters
}
