% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_data.R
\name{get_gadm_code}
\alias{get_gadm_code}
\title{Get GADM code for a particular site}
\usage{
get_gadm_code(site, name_2 = NULL)
}
\arguments{
\item{site}{Site object of admin 1 region}

\item{name_2}{Name of admin 2 region within site. If Null, admin 1 returned.}
}
\description{
Get GADM code for a particular site
}
