% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_mcmc_input}
\alias{get_mcmc_input}
\title{Get MCMC input manual}
\usage{
get_mcmc_input(
  prior,
  likelihood,
  sampler,
  initial,
  n_steps = 100,
  n_chains = 2
)
}
\arguments{
\item{prior}{Monty_model object which outputs prior density}

\item{likelihood}{Monty_model object which outputs likelihood}

\item{sampler}{Monty sampler object}

\item{initial}{Vector of initial parameter values (in same order as prior and likelihood)}

\item{n_steps}{Number of MCMC iterations}

\item{n_chains}{Number of parallel MCMC chains}
}
\description{
A flexible function which takes a pre-defined prior and likelihood and creates a list for use in mcmc_go()
}
