% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_production.R
\name{get_traceplot}
\alias{get_traceplot}
\title{Get MCMC traceplot}
\usage{
get_traceplot(
  pars_array,
  pars = NULL,
  title = "",
  burn_in = 0,
  true_vals_df = NULL,
  save_file = NULL,
  height = 20,
  width = 30
)
}
\arguments{
\item{pars_array}{3d array of MCMC samples in the format [variable, iteration, chain]}

\item{pars}{Names of parameters to show. Default = all parameters}

\item{title}{Plot title}

\item{burn_in}{Length of 'burn-in' to discard from plot}

\item{true_vals_df}{Dataframe with two named columns: variable and value}

\item{save_file}{File location for ggsave}

\item{height}{Height of saved plot in cm}

\item{width}{Width of saved plot in cm}
}
\description{
Get MCMC traceplot
}
