% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mcmc_go}
\alias{mcmc_go}
\title{Run MCMC}
\usage{
mcmc_go(input_list)
}
\arguments{
\item{input_list}{Output from get_mcmc_input()}
}
\description{
Wrapper function for monty::monty_sample. Having a single argument simplifies use of bulk cluster requests.
}
