% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_leap_days}
\alias{n_leap_days}
\title{Count leap days between two dates}
\usage{
n_leap_days(date1, date2)
}
\arguments{
\item{date1}{Date object}

\item{date2}{Date object}
}
\description{
Count leap days between two dates
}
