% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_production.R
\name{restitch_array}
\alias{restitch_array}
\title{Restitch Array}
\usage{
restitch_array(samples_list, var = "pars")
}
\arguments{
\item{samples_list}{List of MCMC outputs from monty::monty_sample or malariafit::mcmc_go}

\item{var}{Variable of interest. Either "pars" or "density".}
}
\description{
This function will take a list of parallel MCMC outputs, and combine the array items into a single array.
Useful for viewing results of multiple chains if you have performed parallel MCMC chains as separate simulations.
}
