% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_bednets_site}
\alias{set_bednets_site}
\title{Set site-specific ITN interventions}
\usage{
set_bednets_site(params, interventions)
}
\arguments{
\item{params}{malariasimple parameters}

\item{interventions}{'interventions' item from site object}
}
\description{
Set site-specific ITN interventions
}
