% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_equilibrium_site}
\alias{set_equilibrium_site}
\title{Set site-specific equilibrium}
\usage{
set_equilibrium_site(params, site)
}
\arguments{
\item{params}{malariasimple parameters}

\item{site}{site object}
}
\description{
Set site-specific equilibrium
}
