% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_interventions_site}
\alias{set_interventions_site}
\title{Set site-specific SMC and ITN interventions}
\usage{
set_interventions_site(params, site)
}
\arguments{
\item{params}{malariasimple parameters}

\item{site}{site object}
}
\description{
Helper function for malariasimple input parameter list.
Takes as input current malariasimple parameters and a site object.
Returns updated malariasimple parameters
}
