% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_rainfall_site}
\alias{set_rainfall_site}
\title{Set site-specific manual carrying capacity input}
\usage{
set_rainfall_site(
  params,
  rainfall_df,
  model = "white",
  tau = NULL,
  zsat = NULL,
  zmax = NULL
)
}
\arguments{
\item{params}{malariasimple parameters}

\item{rainfall_df}{DataFrame containing columns 'Date' and 'Rainfall'. Obtained using get_climate_data(..data_type = "precipitation")}

\item{model}{Rainfall model to be used. Either "white" or "icl"}

\item{tau}{Both models require a specification for tau}

\item{zsat}{'icl' model requires a value for zsat}
}
\description{
Set site-specific manual carrying capacity input
}
