% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_seasonality_site}
\alias{set_seasonality_site}
\title{Set site-specific seasonality}
\usage{
set_seasonality_site(params, site)
}
\arguments{
\item{params}{malariasimple parameters}

\item{site}{site object}
}
\description{
Set site-specific seasonality
}
