% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_wrappers.R
\name{set_smc_site}
\alias{set_smc_site}
\title{Set site-specific SMC interventions}
\usage{
set_smc_site(params, interventions, seasonality)
}
\arguments{
\item{params}{malariasimple parameters}

\item{interventions}{'interventions' item from site object}

\item{seasonality}{'seasonality' item from site object}
}
\description{
Set site-specific SMC interventions
}
