% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_date}
\alias{to_date}
\title{Check whether a date has been entered correctly}
\usage{
to_date(x, arg_name)
}
\arguments{
\item{x}{The date item to check}

\item{arg_name}{the}
}
\description{
Check whether a date has been entered correctly
}
